+++ src/Bundle/CollectionFieldBundle/Resources/webpack/vue/field/CollectionRow.vue
@@ -1,12 +1,12 @@
 <template>
     <div>
         <div class="collection-add-button-wrapper">
-            <button  class="uk-button uk-button-default" v-on:click.prevent="addRowBefore" v-html="feather.icons['plus'].toSvg({ width: 20, height: 20 })"></button>
+            <a href="#" class="uk-button uk-button-default" v-on:click.prevent="addRowBefore" v-html="feather.icons['plus'].toSvg({ width: 20, height: 20 })"></a>
         </div>
         <div class="uk-placeholder uk-padding-small">
             <div class="uk-sortable-handle" v-html="feather.icons['menu'].toSvg({ width: 16, height: 16 })"></div>
             <div :class="formLayout" v-if="prototype" v-html="prototype"></div>
-            <button class="close-button" v-html="feather.icons['x'].toSvg({ width: 20, height: 20 })" v-on:click.prevent="removeRow"></button>
+            <a href="#" class="close-button" v-html="feather.icons['x'].toSvg({ width: 20, height: 20 })" v-on:click.prevent="removeRow"></a>
         </div>
     </div>
 </template>